Moved all dummy classes to one file and namespace
